refactor cli publish into yargs commands
