+++ apps/franklin-bulk/cmds/live.js
@@ -1,0 +1,16 @@
+#!/usr/bin/env node
+
+// imports
+const { cliWorkerHandler } = require('../src/cliWorkerHandler.js');
+const { defaultCLICmdWithWorkerYargsBuilder } = require('../src/yargs.js');
+
+// constants
+const FRANKLIN_STAGE_LIVE = 'live';
+
+/*
+ * Main
+ */
+
+exports.desc = 'Publish pages to live stage on Franklin';
+exports.builder = defaultCLICmdWithWorkerYargsBuilder;
+exports.handler = cliWorkerHandler.bind(null, 'publish_worker.js', { stage: FRANKLIN_STAGE_LIVE });
